+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove unused bundle wiring and agents toggle reconcilers.
+
 ## [0.39.0] - 2025-12-16
 
 ### Removed